requirements.txt
asgiref==3.4.1
blis==0.2.4
certifi==2021.5.30
charset-normalizer==2.0.8
ChatterBot==1.0.5
chatterbot-corpus==1.2.0
click==8.0.1
cymem==2.0.5
dj-database-url==0.5.0
Django==3.2.7
emoji==1.4.2
Flask==2.0.2
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.3
joblib==1.0.1
MarkupSafe==2.0.1
mathparse==0.1.2
murmurhash==1.0.5
nltk==3.6.2
numpy==1.21.2
packaging==21.0
Pint==0.17
plac==0.9.6
preshed==2.0.1
psycopg2==2.9.1
PyJWT==1.7.1
pymongo==4.0
pyparsing==2.4.7
python-dateutil==2.7.5
pytz==2021.1
PyYAML==3.13
regex==2021.8.28
requests==2.26.0
six==1.16.0
spacy==2.1.9
SQLAlchemy==1.2.19
sqlparse==0.4.2
srsly==1.0.5
thinc==7.0.8
tqdm==4.62.2
twilio==6.63.2
urllib3==1.26.6
wasabi==0.8.2
Werkzeug==2.0.1
whitenoise==5.3.0
